Refactoring restricted for team member user
