+++ client/src/app/core/session/session.service.ts
@@ -7,7 +7,7 @@ import { DialogImportComponent } from '../dialog-import/dialog-import.component'
 
 // import { environment } from '../../../environments/environment';
 import { UserRole } from '../user-role';
-import { EditorPermissionsService } from 'src/app/editor/permissions/editor-permissions.service';
+import { EditorPermissionsService } from 'src/app/core/permissions/editor-permissions.service';
 import { EditorNavigationService } from 'src/app/editor/navigation/navigation.service';
 import { TeamsService } from 'src/app/teams/teams.service';
